+++ ui/components/app/add-network/add-network.js
@@ -256,7 +256,7 @@ const AddNetwork = () => {
                   <Button
                     type="inline"
                     className="add-network__add-button"
-                    data-testid={`add-network__${item.nickname}`}
+                    data-testid={`add-network-${item.nickname}`}
                     onClick={async () => {
                       await dispatch(
                         requestUserApproval({
